requirements.txt
-e svn+https://code.djangoproject.com/svn/django/trunk/#egg=django
south==0.7.3
django-autofixture==0.2.2
django-tagging==0.3.1
django-markup==0.4
Markdown==2.1.1
